chore: Update lint and type workflow and commit-msg script
This commit updates the lint and type workflow in the .github/workflows/lint-and-type.yml file. It renames the uild job to lint and type and makes necessary changes to the permissions. Additionally, the commit-msg script in the .husky/commit-msg file is updated to use the command pnpm lint:commitlint instead of pnpm test:commitlint.